requirements.txt
flask
flask-login
flask-script
gunicorn
psycopg2
python-dotenv
